requirements.txt
Django==3.1.1
django-allauth==0.42.0
django-google-analytics==2
django-rest-auth==0.9.5
djangorestframework==3.12.1
google-analytics==0.0.0
gunicorn==20.0.4
Pillow==8.1.0
psycopg2-binary==2.8.6
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0